Refactor rag mcq code
