+++ mcq-rag/generator.py
@@ -1,324 +1,22 @@
-"""
-Enhanced RAG System for Multiple Choice Question Generation
-Author: MathematicGuy
-Date: July 2025
-
-This module implements an advanced RAG system specifically designed for generating
-high-quality Multiple Choice Questions from educational documents.
-"""
-
-import os
+from verify import QualityValidator, DifficultyAnalyzer
+from prompt import PromptTemplateManager
+from util import ContextAwareRetriever, DifficultyLevel, QuestionType, MCQQuestion, MCQOption
+from typing import Dict, Any, Optional, Tuple, List
 import json
-import time
 import torch
-from typing import List, Dict, Any, Optional, Tuple
-from dataclasses import dataclass, asdict
 from pathlib import Path
-from enum import Enum
+import time
 import numpy as np
-
-# LangChain imports
-from langchain_community.document_loaders import PyPDFLoader
 from langchain_huggingface.embeddings import HuggingFaceEmbeddings
-from langchain_experimental.text_splitter import SemanticChunker
 from langchain_huggingface.llms import HuggingFacePipeline
-from langchain_core.output_parsers import JsonOutputParser
-from langchain_core.runnables import RunnableParallel, RunnablePassthrough
-from langchain_core.prompts import PromptTemplate
-from langchain_community.vectorstores import FAISS
-
-from langchain_core.documents import Document
-from unsloth import FastLanguageModel
-
-# Transformers imports
-from transformers import (
-    AutoModelForCausalLM,
-    AutoTokenizer,
-)
-from transformers.pipelines import pipeline
 from transformers.utils.quantization_config import BitsAndBytesConfig
-
-
-class QuestionType(Enum):
-    """Enumeration of different question types"""
-    DEFINITION = "definition"
-    COMPARISON = "comparison"
-    APPLICATION = "application"
-    ANALYSIS = "analysis"
-    EVALUATION = "evaluation"
-
-
-class DifficultyLevel(Enum):
-    """Enumeration of difficulty levels"""
-    EASY = "easy"
-    MEDIUM = "medium"
-    HARD = "hard"
-    EXPERT = "expert"
-
-
-@dataclass
-class MCQOption:
-    """Data class for MCQ options"""
-    label: str
-    text: str
-    is_correct: bool
-
-
-@dataclass
-class MCQQuestion:
-    """Data class for Multiple Choice Question"""
-    question: str
-    context: str
-    options: List[MCQOption]
-    explanation: str
-    difficulty: str
-    topic: str
-    question_type: str
-    source: str
-    confidence_score: float = 0.0
-
-    def to_dict(self) -> Dict[str, Any]:
-        """Convert to dictionary format"""
-        return {
-            "question": self.question,
-            "context": self.context,
-            "options": {opt.label: opt.text for opt in self.options},
-            "correct_answer": next(opt.label for opt in self.options if opt.is_correct),
-            "explanation": self.explanation,
-            "difficulty": self.difficulty,
-            "topic": self.topic,
-            "question_type": self.question_type,
-            "source": self.source,
-            "confidence_score": self.confidence_score
-        }
-
-
-class PromptTemplateManager:
-    """Manages different prompt templates for various question types"""
-
-    def __init__(self):
-        self.base_template = self._create_base_template()
-        self.templates = self._initialize_templates()
-
-    def _create_base_template(self) -> str:
-        """Create the base template used by all question types"""
-        return """
-            Hãy tạo 1 câu hỏi trắc nghiệm dựa trên nội dung sau đây.
-
-            Nội dung: {context}
-            Chủ đề: {topic}
-            Mức độ: {difficulty}
-            Loại câu hỏi: {question_type}
-
-            QUAN TRỌNG: Chỉ trả về JSON hợp lệ, không có text bổ sung, luôn trả lời bằng tiếng việt:
-
-            {{
-                "question": "Câu hỏi rõ ràng về {topic}",
-                "options": {{
-                    "A": "Đáp án A",
-                    "B": "Đáp án B",
-                    "C": "Đáp án C",
-                    "D": "Đáp án D"
-                }},
-                "correct_answer": "A",
-                "explanation": "Giải thích tại sao đáp án A đúng",
-                "topic": "{topic}",
-                "difficulty": "{difficulty}",
-                "question_type": "{question_type}"
-            }}
-
-            Trả lời:
-        """
-
-    def _create_question_specific_instruction(self, question_type: str) -> str:
-        """Create specific instructions for each question type"""
-        instructions = {
-            "definition": "Tạo câu hỏi định nghĩa thuật ngữ. Tập trung vào định nghĩa chính xác.",
-            "application": "Tạo câu hỏi ứng dụng thực tế. Bao gồm tình huống cụ thể.",
-            "analysis": "Tạo câu hỏi phân tích code/sơ đồ. Kiểm tra tư duy phản biện.",
-            "comparison": "Tạo câu hỏi so sánh khái niệm. Tập trung vào điểm khác biệt.",
-            "evaluation": "Tạo câu hỏi đánh giá phương pháp. Yêu cầu quyết định dựa trên tiêu chí."
-        }
-        return instructions.get(question_type, "Tạo câu hỏi chất lượng cao.")
-
-    def _initialize_templates(self) -> Dict[str, str]:
-        """Initialize all templates using the base template"""
-        templates = {"base": self.base_template}
-
-        # Create shorter, more direct templates for each question type
-        instructions = {
-            "definition": "Tạo câu hỏi định nghĩa thuật ngữ.",
-            "application": "Tạo câu hỏi ứng dụng thực tế.",
-            "analysis": "Tạo câu hỏi phân tích code/sơ đồ.",
-            "comparison": "Tạo câu hỏi so sánh khái niệm.",
-            "evaluation": "Tạo câu hỏi đánh giá phương pháp."
-        }
-
-        # Add specific templates for each question type
-        for question_type in QuestionType:
-            instruction = instructions.get(question_type.value, "Tạo câu hỏi chất lượng cao.")
-            templates[question_type.value] = f"{instruction}\n\n{self.base_template}"
-
-        return templates
-
-    def get_template(self, question_type: QuestionType = QuestionType.DEFINITION) -> str:
-        """Get prompt template for specific question type"""
-        return self.templates.get(question_type.value, self.templates["base"])
-
-    def update_base_template(self, new_base_template: str):
-        """Update the base template and regenerate all templates"""
-        self.base_template = new_base_template
-        self.templates = self._initialize_templates()
-        print("✅ Base template updated and all templates regenerated")
-
-    def get_template_info(self) -> Dict[str, int]:
-        """Get information about all templates (for debugging)"""
-        return {
-            template_type: len(template)
-            for template_type, template in self.templates.items()
-        }
-
-class QualityValidator:
-    """Validates the quality of generated MCQ questions"""
-
-    def __init__(self):
-        self.min_question_length = 10
-        self.max_question_length = 200
-        self.min_explanation_length = 20
-
-    def validate_mcq(self, mcq: MCQQuestion) -> Tuple[bool, List[str]]:
-        """Validate MCQ and return validation result with issues"""
-        issues = []
-
-        # Check question length
-        if len(mcq.question) < self.min_question_length:
-            issues.append("Question too short")
-        elif len(mcq.question) > self.max_question_length:
-            issues.append("Question too long")
-
-        # Check options count
-        if len(mcq.options) != 4:
-            issues.append("Must have exactly 4 options")
-
-        # Check for single correct answer
-        correct_count = sum(1 for opt in mcq.options if opt.is_correct)
-        if correct_count != 1:
-            issues.append("Must have exactly one correct answer")
-
-        # Check explanation
-        if len(mcq.explanation) < self.min_explanation_length:
-            issues.append("Explanation too short")
-
-        # Check for distinct options
-        option_texts = [opt.text for opt in mcq.options]
-        if len(set(option_texts)) != len(option_texts):
-            issues.append("Options must be distinct")
-
-        return len(issues) == 0, issues
-
-    def calculate_quality_score(self, mcq: MCQQuestion) -> float:
-        """Calculate quality score from 0 to 100"""
-        is_valid, issues = self.validate_mcq(mcq)
-
-        if not is_valid:
-            return 0.0
-
-        # Start with base score
-        score = 70.0
-
-        # Bonus for good explanation length
-        if len(mcq.explanation) > 50:
-            score += 10
-
-        # Bonus for appropriate question length
-        if 20 <= len(mcq.question) <= 100:
-            score += 10
-
-        # Bonus for diverse option lengths (indicates good distractors)
-        option_lengths = [len(opt.text) for opt in mcq.options]
-        if max(option_lengths) - min(option_lengths) < 50:  # Similar lengths
-            score += 10
-
-        return min(score, 100.0)
-
-
-class DifficultyAnalyzer:
-    """Analyzes and adjusts question difficulty"""
-
-    def __init__(self):
-        self.difficulty_keywords = {
-            DifficultyLevel.EASY: ["là gì", "định nghĩa", "ví dụ", "đơn giản"],
-            DifficultyLevel.MEDIUM: ["so sánh", "khác biệt", "ứng dụng", "khi nào"],
-            DifficultyLevel.HARD: ["phân tích", "đánh giá", "tối ưu", "thiết kế"],
-            DifficultyLevel.EXPERT: ["tổng hợp", "sáng tạo", "nghiên cứu", "phát triển"]
-        }
-
-    def assess_difficulty(self, question: str, context: str) -> DifficultyLevel:
-        """Assess question difficulty based on content analysis"""
-        question_lower = question.lower()
-
-        # Count difficulty indicators
-        difficulty_scores = {}
-        for level, keywords in self.difficulty_keywords.items():
-            score = sum(1 for keyword in keywords if keyword in question_lower)
-            difficulty_scores[level] = score
-
-        # Return highest scoring difficulty
-        if not any(difficulty_scores.values()):
-            return DifficultyLevel.MEDIUM
-
-        return max(difficulty_scores.keys(), key=lambda k: difficulty_scores[k])
-
-
-class ContextAwareRetriever:
-    """Enhanced retriever with context awareness and diversity"""
-
-    def __init__(self, vector_db: FAISS, diversity_threshold: float = 0.7):
-        self.vector_db = vector_db
-        self.diversity_threshold = diversity_threshold
-
-    def retrieve_diverse_contexts(self, query: str, k: int = 5) -> List[Document]:
-        """Retrieve documents with semantic diversity"""
-        # Get more candidates than needed
-        candidates = self.vector_db.similarity_search(query, k=k*2)
-
-        if not candidates:
-            return []
-
-        # Select diverse documents
-        selected = [candidates[0]]  # Always include the most relevant
-
-        for candidate in candidates[1:]:
-            if len(selected) >= k:
-                break
-
-            # Check diversity with already selected documents
-            is_diverse = True
-            for selected_doc in selected:
-                similarity = self._calculate_similarity(candidate.page_content,
-                                                        selected_doc.page_content)
-                if similarity > self.diversity_threshold:
-                    is_diverse = False
-                    break
-
-            if is_diverse:
-                selected.append(candidate)
-
-        return selected[:k]
-
-    def _calculate_similarity(self, text1: str, text2: str) -> float:
-        """Calculate text similarity (simplified implementation)"""
-        words1 = set(text1.lower().split())
-        words2 = set(text2.lower().split())
-
-        if not words1 or not words2:
-            return 0.0
-
-        intersection = words1.intersection(words2)
-        union = words1.union(words2)
-
-        return len(intersection) / len(union)
-
+from unsloth import FastLanguageModel
+from transformers.pipelines import pipeline
+from langchain_community.document_loaders import PyPDFLoader
+from langchain_core.documents import Document
+from langchain_community.vectorstores import FAISS
+from langchain_experimental.text_splitter import SemanticChunker
+from langchain_core.prompts import PromptTemplate
 
 class EnhancedRAGMCQGenerator:
     """Enhanced RAG system for MCQ generation"""
@@ -731,107 +429,3 @@ class EnhancedRAGMCQGenerator:
         print(f"  Template sizes:")
         for template_type, size in template_info.items():
             print(f"    {template_type}: {size} characters")
-
-
-def debug_prompt_templates():
-    """Debug function to test prompt template generation"""
-    print("🔍 Testing Prompt Templates:")
-    prompt_manager = PromptTemplateManager()
-
-    for question_type in QuestionType:
-        try:
-            template = prompt_manager.get_template(question_type)
-            print(f"  {question_type.value}: ✅ Template loaded ({len(template)} chars)")
-        except Exception as e:
-            print(f"  {question_type.value}: ❌ Error - {e}")
-
-
-def main():
-    """Main function demonstrating the enhanced RAG MCQ system"""
-    print("🚀 Starting Enhanced RAG MCQ Generation System")
-
-
-    # Test prompt templates first
-    print("\n🧪 Testing prompt templates...")
-    debug_prompt_templates()
-
-    # Initialize system
-    generator = EnhancedRAGMCQGenerator()
-
-    # Check initial state
-    print("\n🔍 Initial system state:")
-    generator.debug_system_state()
-
-    try:
-        generator.initialize_system()
-
-        # Check state after initialization
-        print("\n🔍 Post-initialization state:")
-        generator.debug_system_state()
-
-        # Load documents
-        folder_path = "pdfs"  # Updated path to your PDF folder
-        try:
-            docs, filenames = generator.load_documents(folder_path)
-            num_chunks = generator.build_vector_database(docs)
-            start = time.time() #? Calc generation time
-
-            print(f"⏱️ Loading Time: {time.time() - start:.2f}s") #? Loading document time
-            print(f"📚 System ready with {len(filenames)} files and {num_chunks} chunks")
-
-            # Generate sample MCQs
-            topics = ["Object Oriented Programming", "Malware Reverse Engineering", "IoT"]
-
-            # Single question generation
-            # print("\n🎯 Generating single MCQ...")
-            # mcq = generator.generate_mcq(
-            #     topic=topics[0],
-            #     difficulty=DifficultyLevel.MEDIUM,
-            #     question_type=QuestionType.DEFINITION
-            # )
-
-            # print(f"Question: {mcq.question}")
-            # print(f"Quality Score: {mcq.confidence_score:.1f}")
-
-            # Batch generation
-            n_question = 2
-            print("\n🎯 Generating batch MCQs...")
-            mcqs = generator.generate_batch(
-                topics=topics,
-                count_per_topic=n_question
-            )
-
-            # Export results
-            output_path = "generated_mcqs.json"
-            generator.export_mcqs(mcqs, output_path)
-
-            # Quality summary
-            print(f"Average mcq generation time taken: {((time.time() - start)/n_question)/60:.2f} min")
-            quality_scores = [mcq.confidence_score for mcq in mcqs]
-            print(f"\n📊 Quality Summary:")
-            print(f"Average Quality: {np.mean(quality_scores):.1f}")
-            print(f"Min Quality: {np.min(quality_scores):.1f}")
-            print(f"Max Quality: {np.max(quality_scores):.1f}")
-
-        except FileNotFoundError as e:
-            print(f"❌ Document folder error: {e}")
-            print("💡 Please ensure your PDF files are in the 'pdfs' folder")
-        except Exception as e:
-            print(f"❌ Document processing error: {e}")
-            print("💡 Check your PDF files and folder structure")
-
-    except Exception as e:
-        print(f"❌ System initialization error: {e}")
-        print("💡 Check your dependencies and API keys")
-        generator.debug_system_state()
-
-
-if __name__ == "__main__":
-    # Check system components
-    # generator = EnhancedRAGMCQGenerator()
-    # generator.debug_system_state()
-
-    # # Test templates separately
-    # debug_prompt_templates()
-
-    main()
